Enhancement/ga (#28)
Migration to GitHub actions
- Build and publish builder container images
- Build and publish container sandbox image
- Generate vulnerability scan log for the container sandbox image
- Generate main binaries for Mac OS, Linux and Windows
- Generate ISO images  for the Live CD
- Refactoring of existing scripts for local builds compatibility

+++ ci_cd/live_cd/build.sh
@@ -10,57 +10,87 @@ PROJECTDIR="$(realpath ${SCRIPTDIR}/../../app)"
 APPVERSION=$(awk -F ' = ' '$1 ~ /version/ { gsub(/[\"]/, "", $2); printf("%s",$2) }' ${PROJECTDIR}/entrusted_client/Cargo.toml)
 CPU_ARCHS="amd64 aarch64"
 ENTRUSTED_VERSION="${APPVERSION}"
+CONTAINER_USER_NAME="entrusted"
+CONTAINER_USER_ID="1024"
 
 if [ -n "$1" ]; then
     ENTRUSTED_VERSION=$1
 fi
 
+echo "Install required packages"
+sudo apt update && sudo apt install -y \
+    debootstrap \
+    squashfs-tools \
+    dosfstools \
+    xorriso \
+    mg xz-utils \
+    isolinux \
+    fakeroot \
+    sudo \
+    bash \
+    wget \
+    syslinux-efi \
+    grub-pc-bin \
+    grub-efi-amd64-bin \
+    systemd-container \
+    bzip2 gzip \
+    mtools
+
 for CPU_ARCH in $CPU_ARCHS ; do
-    LINUX_ARTIFACTSDIR="${PROJECTDIR}/../artifacts/entrusted-linux-${CPU_ARCH}-${ENTRUSTED_VERSION}"
+    MY_TMPDIR="${TMPDIR}"
+    test -d "${MY_TMPDIR}" || MY_TMPDIR="/tmp"
+    ENTRUSTED_ROOT_TMPDIR="${MY_TMPDIR}/entrusted-tmpbuild"
+    LIVE_BOOT_DIR="${ENTRUSTED_ROOT_TMPDIR}/entrusted-livecd/live_boot_dir-${CPU_ARCH}-${ENTRUSTED_VERSION}"
+    LIVE_BOOT_TMP_DIR="${ENTRUSTED_ROOT_TMPDIR}/entrusted-livecd/live_boot_tmpdir-${CPU_ARCH}-${ENTRUSTED_VERSION}"
+    LINUX_ARTIFACTSDIR="${ENTRUSTED_ROOT_TMPDIR}/entrusted-livecd/linux_artifacts-${CPU_ARCH}-${ENTRUSTED_VERSION}"
+    LIVE_ISO_DIR="${PROJECTDIR}/../artifacts"
     DEBIAN_ARCH="amd64"
+    RUST_MUSL_TARGET="x86_64-unknown-linux-musl"
+    RUST_PREAMBLE="RUSTFLAGS='-C target-feature=+crt-static'"
+    test -d "${ENTRUSTED_ROOT_TMPDIR}" && sudo rm -rf "${ENTRUSTED_ROOT_TMPDIR}"
+    mkdir -p "${ENTRUSTED_ROOT_TMPDIR}" "${LIVE_BOOT_TMP_DIR}" "${LIVE_ISO_DIR}" "${LIVE_BOOT_DIR}" "${LINUX_ARTIFACTSDIR}"
 
     if [ ${CPU_ARCH} != "amd64" ]
     then
         DEBIAN_ARCH="arm64"
-    fi
+        RUST_MUSL_TARGET="aarch64-unknown-linux-musl"
+        RUST_PREAMBLE="CC_AARCH64_UNKNOWN_LINUX_MUSL=musl-gcc CXX_AARCH64_UNKNOWN_LINUX_MUSL=musl-g++ AR_AARCH64_UNKNOWN_LINUX_MUSL=ar CARGO_TARGET_AARCH64_UNKNOWN_LINUX_MUSL_LINKER=musl-gcc RUSTFLAGS='-C link-arg=-lgcc -C target-feature=+crt-static'"
+    fi        
     
     echo "Building entrusted version: ${ENTRUSTED_VERSION}"
 
     echo "Cleanup software components build folders"
-    rm -rf ${PROJECTDIR}/entrusted_l10n/target
-    rm -rf ${PROJECTDIR}/entrusted_container/target
-    rm -rf ${PROJECTDIR}/entrusted_client/target
-    rm -rf ${PROJECTDIR}/entrusted_webclient/target
-    rm -rf ${PROJECTDIR}/entrusted_webserver/target
+    test -d ${PROJECTDIR}/entrusted_l10n/target      && rm -rf ${PROJECTDIR}/entrusted_l10n/target
+    test -d ${PROJECTDIR}/entrusted_container/target && rm -rf ${PROJECTDIR}/entrusted_container/target
+    test -d ${PROJECTDIR}/entrusted_client/target    && rm -rf ${PROJECTDIR}/entrusted_client/target
+    test -d ${PROJECTDIR}/entrusted_webclient/target && rm -rf ${PROJECTDIR}/entrusted_webclient/target
+    test -d ${PROJECTDIR}/entrusted_webserver/target && rm -rf ${PROJECTDIR}/entrusted_webserver/target    
+    
+    podman run --platform linux/${DEBIAN_ARCH} -v "${PROJECTDIR}/..":/src -v "${LINUX_ARTIFACTSDIR}":/artifacts docker.io/uycyjnzgntrn/rust-linux:1.64.0 /bin/sh -c "CARGO_NET_GIT_FETCH_WITH_CLI=true CARGO_NET_RETRY=10 ${RUST_PREAMBLE} cargo build --release --target ${RUST_MUSL_TARGET} --manifest-path /src/app/entrusted_webserver/Cargo.toml && cp /src/app/entrusted_webserver/target/${RUST_MUSL_TARGET}/release/entrusted-webserver /artifacts/ && rm -rf /src/app/entrusted_webserver/target && CARGO_NET_GIT_FETCH_WITH_CLI=true CARGO_NET_RETRY=10 ${RUST_PREAMBLE} cargo build --release --target ${RUST_MUSL_TARGET} --manifest-path /src/app/entrusted_client/Cargo.toml && cp /src/app/entrusted_client/target/${RUST_MUSL_TARGET}/release/entrusted-cli /artifacts/ && rm -rf /src/app/entrusted_client/target && strip /artifacts/entrusted-cli && strip /artifacts/entrusted-webserver"
 
     ROOT_SCRIPTS_DIR="$(realpath $(dirname "$0"))"
     chmod +x "${ROOT_SCRIPTS_DIR}"/*.sh
 
-    "${ROOT_SCRIPTS_DIR}"/01-pre-chroot-script.sh "${ENTRUSTED_VERSION}" "${DEBIAN_ARCH}" "${LINUX_ARTIFACTSDIR}"
+    "${ROOT_SCRIPTS_DIR}"/01-pre-chroot.sh "${ENTRUSTED_VERSION}" "${DEBIAN_ARCH}" "${DEBIAN_ARCH}" "${LINUX_ARTIFACTSDIR}" "${LIVE_BOOT_DIR}" "${LIVE_BOOT_TMP_DIR}" "${CONTAINER_USER_NAME}" "${CONTAINER_USER_ID}"
     retVal=$?
     if [ $retVal -ne 0 ]; then
-        echo "Failed to prepare build system for ${CPU_ARCH}"
+        echo "Failed to prepare build for ${CPU_ARCH}"
         exit 1
     fi
 
-    "${ROOT_SCRIPTS_DIR}"/02-just-chroot-script.sh "${DEBIAN_ARCH}"
+    "${ROOT_SCRIPTS_DIR}"/02-just-chroot.sh "${LIVE_BOOT_DIR}" "${LIVE_BOOT_TMP_DIR}" "${ENTRUSTED_VERSION}" "${DEBIAN_ARCH}" "${CONTAINER_USER_NAME}" "${CONTAINER_USER_ID}"
     retVal=$?
     if [ $retVal -ne 0 ]; then
-        echo "Failed to build system for ${CPU_ARCH}"
+        echo "Failed to setup filesystem for ${CPU_ARCH}"
         exit 1
     fi
 
-    "${ROOT_SCRIPTS_DIR}"/05-post-chroot-script.sh "${DEBIAN_ARCH}"
+    "${ROOT_SCRIPTS_DIR}"/04-post-chroot.sh "${DEBIAN_ARCH}" "${LIVE_BOOT_DIR}" "${LIVE_ISO_DIR}"
     retVal=$?
     if [ $retVal -ne 0 ]; then
         echo "Failed to create ISO image for ${CPU_ARCH}"
         exit 1
     fi
-
-    "${ROOT_SCRIPTS_DIR}"/06-upload-iso-image.sh "${DEBIAN_ARCH}" "${CPU_ARCH}"
-    retVal=$?
-    if [ $retVal -ne 0 ]; then
-        echo "Failed to upload iso image for ${CPU_ARCH}"
-        exit 1
-    fi
+    
+    sudo rm -rf "${ENTRUSTED_ROOT_TMPDIR}"
 done
